Extract OpenAPI sorter
